Rename githubUtils to uppercase
